+++ plugins/sigma.layout.forceAtlas2/sigma.layout.forceAtlas2.js
@@ -978,13 +978,16 @@
     }
   };
 
-  sigma.prototype.startForceAtlas2 = function() {
+  sigma.prototype.startForceAtlas2 = function(options) {
     if ((this.forceatlas2 || {}).isRunning)
       return this;
 
     if (!this.forceatlas2) {
-      this.forceatlas2 = new forceatlas2.ForceAtlas2(this.graph);
-      this.forceatlas2.setAutoSettings();
+      this.forceatlas2 = new forceatlas2.ForceAtlas2(this.graph, options || {});
+
+      if (this.forceatlas2.p.autoSettings)
+        this.forceatlas2.setAutoSettings();
+
       this.forceatlas2.init();
     }
 
@@ -1011,9 +1014,6 @@
   };
 
   sigma.prototype.stopForceAtlas2 = function() {
-    if (conrad.hasJob('forceatlas2_' + this.id))
-      conrad.killJob('forceatlas2_' + this.id);
-
     if ((this.forceatlas2 || {}).isRunning) {
       this.forceatlas2.state = {
         step: 0,
@@ -1022,6 +1022,9 @@
       this.forceatlas2.isRunning = false;
       this.forceatlas2.clean();
     }
+
+    if (conrad.hasJob('forceatlas2_' + this.id))
+      conrad.killJob('forceatlas2_' + this.id);
 
     return this;
   };