+++ src/classes/sigma.classes.dispatcher.js
@@ -26,6 +26,7 @@
    */
   dispatcher.prototype.bind = function(events, handler) {
     var i,
+        l,
         event,
         eArray;
 
@@ -41,7 +42,7 @@
     ) {
       eArray = typeof events === 'string' ? events.split(' ') : events;
 
-      for (i in eArray) {
+      for (i = 0, l = eArray.length; i !== l; i += 1) {
         event = eArray[i];
 
         // Check that event is not '':
@@ -76,23 +77,26 @@
    */
   dispatcher.prototype.unbind = function(events, handler) {
     var i,
+        n,
         j,
+        m,
+        k,
         a,
         event,
         eArray = typeof events === 'string' ? events.split(' ') : events;
 
     if (!arguments.length) {
-      for (i in this._handlers)
-        delete this._handlers[i];
+      for (k in this._handlers)
+        delete this._handlers[k];
       return this;
     }
 
     if (handler) {
-      for (i in eArray) {
+      for (i = 0, n = eArray.length; i !== n; i += 1) {
         event = eArray[i];
         if (this._handlers[event]) {
           a = [];
-          for (j in this._handlers[event])
+          for (j = 0, m = this._handlers[event].length; j !== m; j += 1)
             if (this._handlers[event][j].handler !== handler)
               a.push(this._handlers[event][j]);
 
@@ -103,7 +107,7 @@
           delete this._handlers[event];
       }
     } else
-      for (i in eArray)
+      for (i = 0, n = eArray.length; i !== n; i += 1)
         delete this._handlers[eArray[i]];
 
     return this;
@@ -119,7 +123,9 @@
    */
   dispatcher.prototype.dispatchEvent = function(events, data) {
     var i,
+        n,
         j,
+        m,
         a,
         event,
         eventName,
@@ -128,14 +134,14 @@
 
     data = data === undefined ? {} : data;
 
-    for (i in eArray) {
+    for (i = 0, n = eArray.length; i !== n; i += 1) {
       eventName = eArray[i];
 
       if (this._handlers[eventName]) {
         event = self.getEvent(eventName, data);
         a = [];
 
-        for (j in this._handlers[eventName]) {
+        for (j = 0, m = this._handlers[eventName].length; j !== m; j += 1) {
           this._handlers[eventName][j].handler(event);
           if (!this._handlers[eventName][j].one)
             a.push(this._handlers[eventName][j]);